r2r

Enum Error

pub enum Error {

    RCL_RET_OK,
    RCL_RET_ERROR,
    RCL_RET_TIMEOUT,
    RCL_RET_BAD_ALLOC,
    RCL_RET_INVALID_ARGUMENT,
    RCL_RET_UNSUPPORTED,
    RCL_RET_ALREADY_INIT,
    RCL_RET_NOT_INIT,
    RCL_RET_MISMATCHED_RMW_ID,
    RCL_RET_TOPIC_NAME_INVALID,
    RCL_RET_SERVICE_NAME_INVALID,
    RCL_RET_UNKNOWN_SUBSTITUTION,
    RCL_RET_ALREADY_SHUTDOWN,
    RCL_RET_NODE_INVALID,
    RCL_RET_NODE_INVALID_NAME,
    RCL_RET_NODE_INVALID_NAMESPACE,
    RCL_RET_PUBLISHER_INVALID,
    RCL_RET_SUBSCRIPTION_INVALID,
    RCL_RET_SUBSCRIPTION_TAKE_FAILED,
    RCL_RET_CLIENT_INVALID,
    RCL_RET_CLIENT_TAKE_FAILED,
    RCL_RET_SERVICE_INVALID,
    RCL_RET_SERVICE_TAKE_FAILED,
    RCL_RET_TIMER_INVALID,
    RCL_RET_TIMER_CANCELED,
    RCL_RET_WAIT_SET_INVALID,
    RCL_RET_WAIT_SET_EMPTY,
    RCL_RET_WAIT_SET_FULL,
    RCL_RET_INVALID_REMAP_RULE,
    RCL_RET_WRONG_LEXEME,
    RCL_RET_INVALID_PARAM_RULE,
    RCL_RET_INVALID_LOG_LEVEL_RULE,
    RCL_RET_EVENT_INVALID,
    RCL_RET_EVENT_TAKE_FAILED,
    ClockTypeNotRosTime,
    InvalidMessageType {
        msgtype: String,
    },
    SerdeError {
        err: String,
    },
    RCL_RET_ACTION_NAME_INVALID,
    RCL_RET_ACTION_GOAL_ACCEPTED,
    RCL_RET_ACTION_GOAL_REJECTED,
    RCL_RET_ACTION_CLIENT_INVALID,
    RCL_RET_ACTION_CLIENT_TAKE_FAILED,
    RCL_RET_ACTION_SERVER_INVALID,
    RCL_RET_ACTION_SERVER_TAKE_FAILED,
    RCL_RET_ACTION_GOAL_HANDLE_INVALID,
    RCL_RET_ACTION_GOAL_EVENT_INVALID,
    GoalCancelRejected,
    GoalCancelUnknownGoalID,
    GoalCancelAlreadyTerminated,
    InvalidParameterName {
        name: String,
    },
    InvalidParameterType {
        name: String,
        ty: &'static str,
    },
    ParameterValueConv {
        name: String,
        msg: String,
    },
    ParameterWrongType {
        name: String,
        expected_type: &'static str,
        actual_type: &'static str,
    },

}

r2r Error type.

These values are mostly copied straight from the RCL headers, but some are specific to r2r, such as GoalCancelRejected which does not have an analogue in the rcl.

Variants

RCL_RET_OK

RCL_RET_ERROR

RCL_RET_TIMEOUT

RCL_RET_BAD_ALLOC

RCL_RET_INVALID_ARGUMENT

RCL_RET_UNSUPPORTED

RCL_RET_ALREADY_INIT

RCL_RET_NOT_INIT

RCL_RET_MISMATCHED_RMW_ID

RCL_RET_TOPIC_NAME_INVALID

RCL_RET_SERVICE_NAME_INVALID

RCL_RET_UNKNOWN_SUBSTITUTION

RCL_RET_ALREADY_SHUTDOWN

RCL_RET_NODE_INVALID

RCL_RET_NODE_INVALID_NAME

RCL_RET_NODE_INVALID_NAMESPACE

RCL_RET_PUBLISHER_INVALID

RCL_RET_SUBSCRIPTION_INVALID

RCL_RET_SUBSCRIPTION_TAKE_FAILED

RCL_RET_CLIENT_INVALID

RCL_RET_CLIENT_TAKE_FAILED

RCL_RET_SERVICE_INVALID

RCL_RET_SERVICE_TAKE_FAILED

RCL_RET_TIMER_INVALID

RCL_RET_TIMER_CANCELED

RCL_RET_WAIT_SET_INVALID

RCL_RET_WAIT_SET_EMPTY

RCL_RET_WAIT_SET_FULL

RCL_RET_INVALID_REMAP_RULE

RCL_RET_WRONG_LEXEME

RCL_RET_INVALID_PARAM_RULE

RCL_RET_INVALID_LOG_LEVEL_RULE

RCL_RET_EVENT_INVALID

RCL_RET_EVENT_TAKE_FAILED

ClockTypeNotRosTime

InvalidMessageType

Fields

msgtype: String

SerdeError

Fields

err: String

RCL_RET_ACTION_NAME_INVALID

RCL_RET_ACTION_GOAL_ACCEPTED

RCL_RET_ACTION_GOAL_REJECTED

RCL_RET_ACTION_CLIENT_INVALID

RCL_RET_ACTION_CLIENT_TAKE_FAILED

RCL_RET_ACTION_SERVER_INVALID

RCL_RET_ACTION_SERVER_TAKE_FAILED

RCL_RET_ACTION_GOAL_HANDLE_INVALID

RCL_RET_ACTION_GOAL_EVENT_INVALID

GoalCancelRejected

GoalCancelUnknownGoalID

GoalCancelAlreadyTerminated

InvalidParameterName

Fields

name: String

InvalidParameterType

Fields

name: String

ty: &'static str

ParameterValueConv

Fields

name: String

msg: String

ParameterWrongType

Fields

name: String

expected_type: &'static str

actual_type: &'static str

Implementations

impl Error

pub fn from_rcl_error(e: i32) -> Self

pub fn update_param_name(self, param_name: &str) -> Error

Internal function used by code derived for the RosParams trait.